requirements.txt
Pillow>=10.0
streamlit>=1.35
folium>=0.17
streamlit-folium>=0.20
pyproj>=3.6
